LSF launcher class

R6 class to launch and manage LSF workers.

Details

See crew_launcher_lsf().

Attribution

The template files at https://github.com/mschubert/clustermq/tree/master/inst informed the development of the crew launcher plugins in crew.cluster, and we would like to thank Michael Schubert for developing clustermq and releasing it under the permissive Apache License 2.0. See the NOTICE and README.md files in the crew.cluster source code for additional attribution.

See also

Super classes

crew::crew_class_launcher -> crew.cluster::crew_class_launcher_cluster -> crew_class_launcher_lsf

Active bindings

lsf_cwd

See crew_launcher_lsf().

lsf_log_output

See crew_launcher_lsf().

lsf_log_error

See crew_launcher_lsf().

lsf_memory_gigabytes_limit

See crew_launcher_lsf().

lsf_memory_gigabytes_required

See crew_launcher_lsf().

lsf_cores

See crew_launcher_lsf().

Methods

Public methods

crew_class_launcher_lsf$new()

crew_class_launcher_lsf$validate()

crew_class_launcher_lsf$script()

Inherited methods

crew::crew_class_launcher$active()

crew::crew_class_launcher$booting()

crew::crew_class_launcher$call()

crew::crew_class_launcher$done()

crew::crew_class_launcher$errors()

crew::crew_class_launcher$forward()

crew::crew_class_launcher$launch()

crew::crew_class_launcher$rotate()

crew::crew_class_launcher$scale()

crew::crew_class_launcher$set_name()

crew::crew_class_launcher$settings()

crew::crew_class_launcher$start()

crew::crew_class_launcher$summary()

crew::crew_class_launcher$tally()

crew::crew_class_launcher$terminate()

crew::crew_class_launcher$terminate_workers()

crew::crew_class_launcher$unlaunched()

crew::crew_class_launcher$wait()

crew.cluster::crew_class_launcher_cluster$launch_worker()

crew.cluster::crew_class_launcher_cluster$terminate_worker()

Method new()

LSF launcher constructor.

Usage

crew_class_launcher_lsf$new(
  name = NULL,
  seconds_interval = NULL,
  seconds_timeout = NULL,
  seconds_launch = NULL,
  seconds_idle = NULL,
  seconds_wall = NULL,
  tasks_max = NULL,
  tasks_timers = NULL,
  reset_globals = NULL,
  reset_packages = NULL,
  reset_options = NULL,
  garbage_collection = NULL,
  launch_max = NULL,
  tls = NULL,
  verbose = NULL,
  command_submit = NULL,
  command_terminate = NULL,
  script_directory = NULL,
  script_lines = NULL,
  lsf_cwd = NULL,
  lsf_log_output = NULL,
  lsf_log_error = NULL,
  lsf_memory_gigabytes_limit = NULL,
  lsf_memory_gigabytes_required = NULL,
  lsf_cores = NULL
)

Arguments

name

See crew_launcher_lsf().

seconds_interval

See crew_launcher_lsf().

seconds_timeout

See crew_launcher_lsf().

seconds_launch

See crew_launcher_lsf().

seconds_idle

See crew_launcher_lsf().

seconds_wall

See crew_launcher_lsf().

tasks_max

See crew_launcher_lsf().

tasks_timers

See crew_launcher_lsf().

reset_globals

See crew_launcher_lsf().

reset_packages

See crew_launcher_lsf().

reset_options

See crew_launcher_lsf().

garbage_collection

See crew_launcher_lsf().

launch_max

See crew_launcher_lsf().

tls

See crew_launcher_lsf().

verbose

See crew_launcher_lsf().

command_submit

See crew_launcher_lsf().

command_terminate

See crew_launcher_lsf().

script_directory

See crew_launcher_lsf().

script_lines

See crew_launcher_lsf().

lsf_cwd

See crew_launcher_lsf().

lsf_log_output

See crew_launcher_lsf().

lsf_log_error

See crew_launcher_lsf().

lsf_memory_gigabytes_limit

See crew_launcher_lsf().

lsf_memory_gigabytes_required

See crew_launcher_lsf().

lsf_cores

See crew_launcher_lsf().

Returns

an LSF launcher object.

Method validate()

Validate the launcher.

Usage

crew_class_launcher_lsf$validate()

Returns

NULL (invisibly). Throws an error if a field is invalid.

Method script()

Generate the job script.

Usage

crew_class_launcher_lsf$script(name)

Arguments

name

Character of length 1, name of the job. For inspection purposes, you can supply a mock job name.

Details

Includes everything except the worker-instance-specific job name and the worker-instance-specific call to crew::crew_worker(), both of which get inserted at the bottom of the script at launch time.

Returns

Character vector of the lines of the job script.

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
launcher <- crew_launcher_lsf(
  lsf_cwd = getwd(),
  lsf_log_output = "log_file_%J.log",
  lsf_log_error = NULL,
  lsf_memory_gigabytes_limit = 4
)
launcher$script(name = "my_job_name")
}

Examples

## ------------------------------------------------
## Method `crew_class_launcher_lsf$script`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
launcher <- crew_launcher_lsf(
  lsf_cwd = getwd(),
  lsf_log_output = "log_file_%J.log",
  lsf_log_error = NULL,
  lsf_memory_gigabytes_limit = 4
)
launcher$script(name = "my_job_name")
}