PBS/TORQUE launcher class

R6 class to launch and manage PBS/TORQUE workers.

Details

See crew_launcher_pbs().

Attribution

The template files at https://github.com/mschubert/clustermq/tree/master/inst informed the development of the crew launcher plugins in crew.cluster, and we would like to thank Michael Schubert for developing clustermq and releasing it under the permissive Apache License 2.0. See the NOTICE and README.md files in the crew.cluster source code for additional attribution.

See also

Super classes

crew::crew_class_launcher -> crew.cluster::crew_class_launcher_cluster -> crew_class_launcher_pbs

Methods

Public methods

crew_class_launcher_pbs$validate()

crew_class_launcher_pbs$script()

Inherited methods

crew::crew_class_launcher$active()

crew::crew_class_launcher$booting()

crew::crew_class_launcher$call()

crew::crew_class_launcher$done()

crew::crew_class_launcher$errors()

crew::crew_class_launcher$forward()

crew::crew_class_launcher$launch()

crew::crew_class_launcher$rotate()

crew::crew_class_launcher$scale()

crew::crew_class_launcher$set_name()

crew::crew_class_launcher$settings()

crew::crew_class_launcher$start()

crew::crew_class_launcher$summary()

crew::crew_class_launcher$tally()

crew::crew_class_launcher$terminate()

crew::crew_class_launcher$terminate_workers()

crew::crew_class_launcher$unlaunched()

crew::crew_class_launcher$wait()

crew.cluster::crew_class_launcher_cluster$initialize()

crew.cluster::crew_class_launcher_cluster$launch_worker()

crew.cluster::crew_class_launcher_cluster$terminate_worker()

Method validate()

Validate the launcher.

Usage

crew_class_launcher_pbs$validate()

Returns

NULL (invisibly). Throws an error if a field is invalid.

Method script()

Generate the job script.

Usage

crew_class_launcher_pbs$script(name)

Arguments

name

Character of length 1, name of the job. For inspection purposes, you can supply a mock job name.

Details

Includes everything except the worker-instance-specific job name and the worker-instance-specific call to crew::crew_worker(), both of which get inserted at the bottom of the script at launch time.

Returns

Character vector of the lines of the job script.

Examples

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
launcher <- crew_launcher_pbs(
  pbs_cores = 2,
  pbs_memory_gigabytes_required = 4
)
launcher$script(name = "my_job_name")
}

Examples

## ------------------------------------------------
## Method `crew_class_launcher_pbs$script`
## ------------------------------------------------

if (identical(Sys.getenv("CREW_EXAMPLES"), "true")) {
launcher <- crew_launcher_pbs(
  pbs_cores = 2,
  pbs_memory_gigabytes_required = 4
)
launcher$script(name = "my_job_name")
}