SGE launcher class

R6 class to launch and manage SGE workers.

Details

See crew_launcher_sge().

See also

Super classes

crew::crew_class_launcher -> crew.cluster::crew_class_launcher_cluster -> crew_class_launcher_sge

Public fields

sge_cwd

See crew_launcher_sge().

sge_envvars

See crew_launcher_sge().

sge_log_output

See crew_launcher_sge().

sge_log_error

See crew_launcher_sge().

sge_log_join

See crew_launcher_sge().

sge_memory_gigabytes_required

See crew_launcher_sge().

sge_memory_gigabytes_limit

See crew_launcher_sge().

sge_cores

See crew_launcher_sge().

sge_gpu

See crew_launcher_sge().

Methods

Public methods

crew_class_launcher_sge$new()

crew_class_launcher_sge$validate()

crew_class_launcher_sge$script()

Inherited methods

crew::crew_class_launcher$call()

crew::crew_class_launcher$expected()

crew::crew_class_launcher$launch()

crew::crew_class_launcher$launching()

crew::crew_class_launcher$settings()

crew::crew_class_launcher$start()

crew::crew_class_launcher$terminate()

crew.cluster::crew_class_launcher_cluster$args_launch()

crew.cluster::crew_class_launcher_cluster$args_terminate()

crew.cluster::crew_class_launcher_cluster$launch_worker()

crew.cluster::crew_class_launcher_cluster$terminate_worker()

Method new()

SGE launcher constructor.

Usage

crew_class_launcher_sge$new(
  name = NULL,
  seconds_launch = NULL,
  seconds_interval = NULL,
  seconds_timeout = NULL,
  seconds_idle = NULL,
  seconds_wall = NULL,
  seconds_exit = NULL,
  tasks_max = NULL,
  tasks_timers = NULL,
  reset_globals = NULL,
  reset_packages = NULL,
  reset_options = NULL,
  garbage_collection = NULL,
  verbose = NULL,
  command_submit = NULL,
  command_delete = NULL,
  script_directory = NULL,
  script_lines = NULL,
  sge_cwd = NULL,
  sge_envvars = NULL,
  sge_log_output = NULL,
  sge_log_error = NULL,
  sge_log_join = NULL,
  sge_memory_gigabytes_required = NULL,
  sge_memory_gigabytes_limit = NULL,
  sge_cores = NULL,
  sge_gpu = NULL
)

Arguments

name

See crew_launcher_sge().

seconds_launch

See crew_launcher_sge().

seconds_interval

See crew_launcher_sge().

seconds_timeout

See crew_launcher_sge().

seconds_idle

See crew_launcher_sge().

seconds_wall

See crew_launcher_sge().

seconds_exit

See crew_launcher_sge().

tasks_max

See crew_launcher_sge().

tasks_timers

See crew_launcher_sge().

reset_globals

See crew_launcher_sge().

reset_packages

See crew_launcher_sge().

reset_options

See crew_launcher_sge().

garbage_collection

See crew_launcher_sge().

verbose

See crew_launcher_sge().

command_submit

See crew_launcher_sge().

command_delete

See crew_launcher_sge().

script_directory

See crew_launcher_sge().

script_lines

See crew_launcher_sge().

sge_cwd

See crew_launcher_sge().

sge_envvars

See crew_launcher_sge().

sge_log_output

See crew_launcher_sge().

sge_log_error

See crew_launcher_sge().

sge_log_join

See crew_launcher_sge().

sge_memory_gigabytes_required

See crew_launcher_sge().

sge_memory_gigabytes_limit

See crew_launcher_sge().

sge_cores

See crew_launcher_sge().

sge_gpu

See crew_launcher_sge().

Returns

an SGE launcher object.

Method validate()

Validate the launcher.

Usage

crew_class_launcher_sge$validate()

Returns

NULL (invisibly). Throws an error if a field is invalid.

Method script()

Generate the job script.

Usage

crew_class_launcher_sge$script(name)

Arguments

name

Character of length 1, name of the job. For inspection purposes, you can supply a mock job name.

Details

Includes everything except the worker-instance-specific job name and the worker-instance-specific call to crew::crew_worker(), both of which get inserted at the bottom of the script at launch time.

Returns

Character vector of the lines of the job script.

Examples

launcher <- crew_launcher_sge(
  sge_cores = 2,
  sge_memory_gigabytes_required = 4
)
launcher$script(name = "my_job_name")

Examples

## ------------------------------------------------
## Method `crew_class_launcher_sge$script`
## ------------------------------------------------

launcher <- crew_launcher_sge(
  sge_cores = 2,
  sge_memory_gigabytes_required = 4
)
launcher$script(name = "my_job_name")
#> [1] "#$ -N my_job_name"   "#$ -cwd"             "#$ -o /dev/null"    
#> [4] "#$ -j y"             "#$ -l m_mem_free=4G" "#$ -pe smp 2"